+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * add conditionals for bean availability and system properties
 * add @LoggedIn annotation for SecurityContext assertions
 * add abstract server-side WebSocketAdapter with JSON support
+* add AWS Cloud tooling with @AWSClient autowiring
 
 # Version 1.3
 * Update 3rd-party libraries